style: apply repo-wide biome rewrites
Capture the repo-wide Biome formatting pass and fold the tiny BMP palette helpers back into render-bmp.ts so the monochrome palette fix stays self-contained.

Made-with: Cursor
