+++ app/(render)/recipes/[slug]/preview/layout.tsx
@@ -1,4 +1,4 @@
-import { Suspense, type ReactNode } from "react";
+import { type ReactNode, Suspense } from "react";
 
 export default function RecipePreviewLayout({
 	children,
